Update xmr.cpp
This code replaces the default rx_slow_hash function with a version optimized for the AVX instruction set, which is available on modern CPUs.

    Optimize memory allocation: Reduce the number of memory allocations by reusing memory whenever possible. This can be achieved by using memory pools or by pre-allocating memory buffers for frequently used data structures. Here is an example of pre-allocating memory for the hashing output buffer in get_hashing_blob function:

This code pre-allocates a static buffer hash_buffer of size HASH_SIZE for the hashing output, and reuses it for each call to get_hashing_blob. This eliminates the need for allocating and deallocating memory for the output buffer on each call. Note that this assumes the output size is always HASH_SIZE.