Update xmr.cpp
Use SIMD instructions: Modern CPUs come with support for SIMD (Single Instruction Multiple Data) instructions, which allow multiple calculations to be performed on multiple data elements at the same time. In the case of Monero mining, this can result in a significant speedup. The Monero-Pool code already has an implementation for utilizing SIMD instructions using the crypto/keccak.h and crypto/slow-hash.c files. To utilize SIMD instructions in the get_rx_hash function, you can replace the call to rx_slow_hash with a call to keccakf1600, like so:

This will use the keccakf1600 function, which utilizes SIMD instructions to perform the hash calculation.